+++ src/main/java/com/project/doubleshop/domain/cart/entity/Cart.java
@@ -1,5 +1,7 @@
 package com.project.doubleshop.domain.cart.entity;
 
+import java.util.List;
+
 import javax.persistence.Entity;
 import javax.persistence.FetchType;
 import javax.persistence.GeneratedValue;
