+++ src/main/java/com/project/doubleshop/domain/order/repository/OrderDetailRepository.java
@@ -1,0 +1,11 @@
+package com.project.doubleshop.domain.order.repository;
+
+import java.util.List;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import com.project.doubleshop.domain.order.entity.OrderDetail;
+
+public interface OrderDetailRepository extends JpaRepository<OrderDetail, Long> {
+	List<OrderDetail> findByOrderId(Long orderId);
+}
